+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Accept regular formatters after conditional formatters
 
 ### v4.0.0-beta.1
+  - Fix multiple reDoS and optimize parsing of some templates
   - [EE] Include 3.5.2
   - [EE] Dynamic chart: 
     - Fix crash when DOCX/ODT templates contain empty files